+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `starknet_getCompiledCasm` returns CASM wrapped in a `casm` property.
 - `starknet_traceBlockTransactions` fails on Starknet 0.13.4 when a fallback to fetching from the feeder gateway is required.
 - Websocket subscriptions to the `latest` block do not send notifications for the current latest block.
+- `starknet_subscribeEvents` subscriptions send matching events only from the `latest` block, not as soon as those show up in `pending`.
 
 ## [0.16.1] - 2025-02-24